fix: move SearchResultsContainer to shared/ui and use semantic tailwind variables
- Moved UI component from src/app/[locale]/search/ui/ to src/shared/ui/search/
- Replaced hardcoded tailwind colors (text-gray-*, border-gray-*, bg-blue-*)
  with semantic variables (text-text, text-muted, border-border, bg-card, etc.)
- Updated imports in search page to use new component location

Fixes PR review comments:
- Component location: UI components now in shared/ui folder
- Styling: Uses tailwind variables defined in globals.css
